Revert "Avoid dist-upgrade"
This reverts commit 8b54108dda5b6af15c220d6c073990b81010249e.

In that commit I assumed that we would still test with up-to-date
packages every night because gap-docker-master is rebuilt daily.
However, gap-docker-master is based on gap-docker-base which is
only rebuilt now and then. Thus, without dist-upgrade we test an
arbitrary point in time (or even two points in time mixed, one
when gap-docker-base was rebuilt and one when gap-docker(-master)
was built). We want to avoid this situation.